+++ R/pca_cor_samplevar.R
@@ -29,7 +29,7 @@ NULL
 #' pca <- calc_pca(object)$x
 #'
 #' # Obtain the sample variable.
-#' x <- colData(object)$low_depth_flag
+#' x <- colData(object)$AGE18
 #'
 #' # Correlate them.
 #' r2 <- h_pca_var_rsquared(pca, x)