+++ NEWS.md
@@ -6,6 +6,8 @@
 * The `multi_assay_experiment` now contains `HermesData` experiments, different patient IDs, one experiment with normalized assays, and multiple samples per patient in one experiment.
 * `calc_pca` now includes a top number of gene filtering option `n_top`, which allows filtering genes with greatest variability across samples. 
 * Generation of gene signatures with the functions `colPrinComp1` and `colMeanZscores` using the PC1 and the Z-score respectively.
+* Renaming of required `rowData` and `colData` columns to be more consistent with standards and use lowercase snake-case names.
+* Annotation querying and setting is now more flexible in that it also allows to query more annotations than the required ones.
 
 ### Bug Fixes
 * `normalize()` now also works when the `hermes` package is not loaded, i.e. you can use it with `hermes::normalize()`.